+++ frontend/src/views/clusterview/ClusterStationGroupsView.vue
@@ -78,6 +78,9 @@ const sortedGroupMembers = computed(() =>
 
 const {
   availableMembers,
+  offeredUserTypes,
+  search: memberSearch,
+  userType: memberUserType,
   addMember: addStationToGroup,
   removeMember: removeStationFromGroup,
 } = useMemberAssignment(
@@ -104,6 +107,8 @@ const {
                         @delete="requestDelete"/>
         <GroupDetailPanel v-if="selectedGroup" :selected-group="selectedGroup" :group-loading="groupLoading"
                           :sorted-group-members="sortedGroupMembers" :available-members="availableMembers"
+                          :offered-user-types="offeredUserTypes"
+                          v-model:member-search="memberSearch" v-model:member-user-type="memberUserType"
                           :group-roles="groupRoles" :all-roles="allRoles" v-model:group-role-ids="groupRoleIds"
                           :can-edit-roles="false" @add-member="addStationToGroup"
                           @remove-member="removeStationFromGroup"/>